requirements.txt
dj-database-url==0.5.0
Django==3.0.7
Jinja2==2.10.1
pytz==2018.5
sqlparse==0.3.1
djangorestframework==3.10.0
djangorestframework-jwt==1.11.0


